Move class declaration to the main feature file
